Add `Rank`, `Suit`, and `Hand` structs
Adds a collection of simple utility structs to make it a bit more clear
what is going on in the code, and replace most hard-coded constants with
named constants associated with a type.

Fix a minor bug in `draw_random_card`, where the upper bound of the
range was actually 50 instead of 51, thus making it impossible to draw
the king of spades.

Run `cargo clippy` on the project and fix all standard lints.