Add `HandScore` type, tests, and utilities (#4)
Adds the new `HandScore` type, which is used to store the "score" of a poker
hand:  for example, does the hand contain a pair? `HandScore` includes a
Boolean flag for every legal combination of cards used in scoring a poker hand.
Rather then picking only the best hand, it records whether the hand satisfies
the requirements for each "score". For example, all hands with a "three of a
kind" will also contain a "pair".

Along with this new data type comes a collection of utility functions and traits
that can be used to "score" (or convert) a hand of cards into a `HandScore`.
These functions are a bit tricky to get right, so I've included a set of unit
tests to ensure that they do the right thing for a selection of representative
hands (for both 5 and 7 card variants of the game).

Finally, there are some other minor clean-ups that made it easier to write the
unit test and maintain the code, such as breaking out `Hand` into its own file
and converting many of the constructors to `From` traits.